refactor: extract ClasspathHelper from package object
has become quite large, and will continue to grow...
